Feature/66 currency

+++ src/main/java/com/moneyme/moneymebackend/dto/request/ObligationRequestDTO.java
@@ -7,6 +7,6 @@ import java.math.BigDecimal;
 
 @Data
 public class ObligationRequestDTO {
-    @JsonProperty("amount") BigDecimal amount;
+    @JsonProperty("amount") Integer amount;
     @JsonProperty("user_uuid") String userUuid ;
 }
